Limpieza de datos de la entidad persona en la BD. Agregado BCrypt para mayor seguridad en login. Actualizado ControladorPersona con nuevo metodo validarLogin Actualizado en persona atributos dni/telefono => long Implementado login. Fix bug de mensaje de bienvenida.

+++ eclipse-workspace/tpJava/src/datos/DatosAdministrador.java
@@ -105,7 +105,7 @@ public class DatosAdministrador {
 			//roundtriptime ida y vuelta hasta la bd
 			stmt.setString(1, p.getNombre());
 			stmt.setString(2, p.getApellido());
-			stmt.setInt(3, p.getDni());
+			stmt.setLong(3, p.getDni());
 			stmt.setLong(4, p.getTelefono());
 			stmt.setString(5, p.getDireccion());
 			stmt.setString(6, p.getEmail());
@@ -140,7 +140,7 @@ public class DatosAdministrador {
 			
 			stmt.setString(2, p.getNombre());
 			stmt.setString(3, p.getApellido());
-			stmt.setInt(4, p.getDni());
+			stmt.setLong(4, p.getDni());
 			stmt.setLong(5, p.getTelefono());
 			stmt.setString(6, p.getDireccion());
 			stmt.setString(7, p.getEmail());
@@ -171,7 +171,7 @@ public class DatosAdministrador {
 			
 			stmt.setString(1, p.getNombre());
 			stmt.setString(2, p.getApellido());
-			stmt.setInt(3, p.getDni());
+			stmt.setLong(3, p.getDni());
 			stmt.setLong(4, p.getTelefono());
 			stmt.setString(5, p.getDireccion());
 			stmt.setString(6, p.getEmail());
